services/grade-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3
PyJWT==2.8.0
zeep==4.2.1